+++ annotation/compiler/src/main/java/com/bumptech/glide/annotation/compiler/AppModuleProcessor.java
@@ -17,7 +17,7 @@ import javax.lang.model.element.TypeElement;
  * Runs the final steps of Glide's annotation process and generates the combined
  * {@link AppGlideModule}, {@link com.bumptech.glide.Glide},
  * {@link com.bumptech.glide.RequestManager}, and
- * {@link com.bumptech.glide.request.BaseRequestOptions} classes.
+ * {@link com.bumptech.glide.request.RequestOptions} classes.
  */
 final class AppModuleProcessor {
   private static final String COMPILER_PACKAGE_NAME =
