+++ annotation/compiler/src/main/java/com/bumptech/glide/annotation/compiler/RequestManagerGenerator.java
@@ -251,7 +251,7 @@ final class RequestManagerGenerator {
   }
 
   /**
-   * The {@link com.bumptech.glide.request.BaseRequestOptions} subclass should always be our
+   * The {@link com.bumptech.glide.request.RequestOptions} subclass should always be our
    * generated subclass type to avoid inadvertent errors where a different subclass is applied that
    * accidentally wipes out some logic in overidden methods in our generated subclass.
    */
@@ -263,9 +263,9 @@ final class RequestManagerGenerator {
     }
 
     Elements elementUtils = processingEnv.getElementUtils();
-    TypeElement baseRequestOptionsType =
+    TypeElement requestOptionsType =
             elementUtils.getTypeElement(
-                RequestOptionsGenerator.BASE_REQUEST_OPTIONS_QUALIFIED_NAME);
+                RequestOptionsGenerator.REQUEST_OPTIONS_QUALIFIED_NAME);
     TypeElement androidNonNullType =
             elementUtils.getTypeElement("android.support.annotation.NonNull");
 
@@ -281,7 +281,7 @@ final class RequestManagerGenerator {
         .addAnnotation(Override.class)
         .addModifiers(Modifier.PROTECTED)
         .addParameter(
-            ParameterSpec.builder(ClassName.get(baseRequestOptionsType), parameterName)
+            ParameterSpec.builder(ClassName.get(requestOptionsType), parameterName)
                 .addAnnotation(ClassName.get(androidNonNullType))
                 .build())
         .beginControlFlow("if ($N instanceof $L)",
