+++ annotation/src/main/java/com/bumptech/glide/annotation/GlideType.java
@@ -22,7 +22,7 @@ import java.lang.annotation.Target;
  * with {@link GlideExtension}.
  *
  * <p>Methods annotated with GlideType must have a single parameter. The type of the
- * single parameter must be {@link com.bumptech.glide.request.BaseRequestOptions}, with a type
+ * single parameter must be {@link com.bumptech.glide.request.RequestOptions}, with a type
  * matching the value of {@link #value()}.
  */
 @Target(ElementType.METHOD)
@@ -33,7 +33,7 @@ public @interface GlideType {
   /**
    * A Resource class name, like GifDrawable.class, Bitmap.class etc.
    *
-   * <p>Must match the type of the {@link com.bumptech.glide.request.BaseRequestOptions} parameter.
+   * <p>Must match the type of the {@link com.bumptech.glide.request.RequestOptions} parameter.
    */
   Class<?> value();
 }
