+++ library/src/main/java/com/bumptech/glide/RequestManager.java
@@ -22,7 +22,6 @@ import com.bumptech.glide.manager.LifecycleListener;
 import com.bumptech.glide.manager.RequestManagerTreeNode;
 import com.bumptech.glide.manager.RequestTracker;
 import com.bumptech.glide.manager.TargetTracker;
-import com.bumptech.glide.request.BaseRequestOptions;
 import com.bumptech.glide.request.Request;
 import com.bumptech.glide.request.RequestOptions;
 import com.bumptech.glide.request.target.Target;
@@ -66,7 +65,7 @@ public class RequestManager implements LifecycleListener {
   private final ConnectivityMonitor connectivityMonitor;
 
   @NonNull
-  private BaseRequestOptions<?> requestOptions;
+  private RequestOptions requestOptions;
 
   public RequestManager(Glide glide, Lifecycle lifecycle, RequestManagerTreeNode treeNode) {
     this(glide, lifecycle, treeNode, new RequestTracker(), glide.getConnectivityMonitorFactory());
@@ -106,11 +105,11 @@ public class RequestManager implements LifecycleListener {
     glide.registerRequestManager(this);
   }
 
-  protected void setRequestOptions(@NonNull BaseRequestOptions<?> toSet) {
+  protected void setRequestOptions(@NonNull RequestOptions toSet) {
     this.requestOptions = toSet.clone().autoClone();
   }
 
-  private void updateRequestOptions(BaseRequestOptions<?> toUpdate) {
+  private void updateRequestOptions(RequestOptions toUpdate) {
     this.requestOptions.apply(toUpdate);
   }
 
@@ -128,7 +127,7 @@ public class RequestManager implements LifecycleListener {
    *
    * <p>The modified options will only be applied to loads started after this method is called.
    *
-   * @see RequestBuilder#apply(BaseRequestOptions)
+   * @see RequestBuilder#apply(RequestOptions)
    *
    * @return This request manager.
    */
@@ -449,7 +448,7 @@ public class RequestManager implements LifecycleListener {
     requestTracker.runRequest(request);
   }
 
-  BaseRequestOptions<?> getDefaultRequestOptions() {
+  RequestOptions getDefaultRequestOptions() {
     return requestOptions;
   }
