+++ library/src/test/java/com/bumptech/glide/RequestBuilderTest.java
@@ -8,7 +8,6 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 import android.widget.ImageView;
-import com.bumptech.glide.request.BaseRequestOptions;
 import com.bumptech.glide.request.Request;
 import com.bumptech.glide.request.RequestOptions;
 import com.bumptech.glide.request.target.Target;
@@ -116,7 +115,7 @@ public class RequestBuilderTest {
         .thenReturn(mock(Target.class));
     when(glideContext.getDefaultRequestOptions()).thenReturn(new RequestOptions());
     when(requestManager.getDefaultRequestOptions())
-        .thenReturn((BaseRequestOptions) new RequestOptions());
+        .thenReturn((RequestOptions) new RequestOptions());
     return new RequestBuilder<>(glide, requestManager, Object.class)
         .load((Object) null);
   }
